requirements.txt
opencv-python~=4.5.1.48
numpy~=1.20.2
sklearn~=0.0
scikit-learn~=0.24.2
fastapi[all]~=0.63.0
python-multipart~=0.0.5
pydantic~=1.8.1
passlib~=1.7.4
python-dotenv~=0.17.1
python-jose[cryptography]==3.2.0